requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
pandas==2.1.4
sqlalchemy==2.0.23
python-dotenv==1.0.0